fix - Считаем прогулы
